fix(src/patch): prevent UnhandledPromiseRejectionWarning in node 7
Starting from node 7, unhandled rejected promises are displaying a
warning to the stdout: `UnhandledPromiseRejectionWarning: Unhandled
promise rejection (rejection id: 1): ...`

On top of that, a message telling that unhandled promise rejections will
terminate the node process in the future.

The unhandled rejection occurs because the `returnValue.then(...)`
statement returns a new rejected promise if the test is failing, but
this returned promise isn't used.

This change makes promise resolution and rejection handling both at the
same place, so no promise should be left unhandled.